2024/day09/main_test.go
package main

import (
	"github.com/stretchr/testify/require"
	"testing"
)

const testInputA = `2333133121414131402
`

const testInputB = testInputA

func TestSolutionA(t *testing.T) {
	type test struct {
		name   string
		input  string
		expect int
	}

	tests := []test{
		{"basic A", testInputA, 1928},
	}

	for _, tt := range tests {
		t.Run(`2024-09 A `+tt.name, func(t *testing.T) {
			result := solutionA([]byte(tt.input))
			require.Equal(t, tt.expect, result)
		})
	}
}

func TestSolutionB(t *testing.T) {
	type test struct {
		name   string
		input  string
		expect int
	}

	tests := []test{
		{"basic B", testInputB, 2858},
	}

	for _, tt := range tests {
		t.Run(`2024-09 B `+tt.name, func(t *testing.T) {
			result := solutionB([]byte(tt.input))
			require.Equal(t, tt.expect, result)
		})
	}
}
